move sort options, card options, and search options to stage level and out of prompt level
