+++ src/components/enhancers/withDisabledAssetRequired.js
@@ -1,0 +1,6 @@
+import { withProps } from 'recompose';
+
+const withDisabledAssetRequired =
+  withProps(({ dataSource }) => ({ disabled: !dataSource }));
+
+export default withDisabledAssetRequired;
