+++ src/components/sections/NameGeneratorAutoCompletePrompts/NameGeneratorAutoCompletePrompts.js
@@ -3,8 +3,10 @@ import { compose } from 'recompose';
 import EditableList from '../../EditableList';
 import withSubject from '../../enhancers/withSubject';
 import withDisabledSubjectRequired from '../../enhancers/withDisabledSubjectRequired';
+import withDisabledAssetRequired from '../../enhancers/withDisabledAssetRequired';
 import { PromptPreview } from '../NameGeneratorPrompts';
 import PromptFields from './PromptFields';
+import withDataSource from '../../enhancers/withDataSource';
 
 const NameGeneratorAutoCompletePrompts = props => (
   <EditableList
@@ -26,5 +28,7 @@ export { NameGeneratorAutoCompletePrompts };
 
 export default compose(
   withSubject,
+  withDataSource,
   withDisabledSubjectRequired,
+  withDisabledAssetRequired,
 )(NameGeneratorAutoCompletePrompts);
